# helpers/swagger.yaml
openapi: 3.0.0
info:
  title: The System MySQL Database
  description: The System MySQL Database
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /outcomes:
    get:
      summary: Get a list of all outcomes
      description: Retrieve list of all outcomes.
      operationId: getAllOutcomes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: An array of all outcomes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 123
                    outcomeStatus:
                      type: string
                      example: 'Open'
                    transactionType:
                      type: string
                      example: 'Call'
                    numberCalled:
                      type: string
                      example: '0123456789'
                    emailUsed:
                      type: string
                      example: 'pluto@example.com'
                    contactPerson:
                      type: string
                      example: 'Pluto'
                    outcomeResolution:
                      type: string
                      example: 'Customer requested a callback'
                    ptpDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    ptpAmount:
                      type: integer
                      example: 1500.00
                    debitResubmissionDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    debitResubmissionAmount:
                      type: integer
                      example: 2000.00
                    outcomeNotes:
                      type: string
                      example: 'Need to call back the customer'
                    nextSteps:
                      type: string
                      example: 'Need to keep pressing for payment'
                    createdDate:
                      type: string
                      example: '2020-05-05T09:12:57.848Z'
                    createdBy:
                      type: string
                      example: 'mickey@example.com'
                    closedDate:
                      type: string
                      example: '2020-05-08T03:11:21.553Z'
                    closedBy:
                      type: string
                      example: 'mickey@example.com'
                    f_caseId:
                      type: integer
                      example: 123
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    post:
      summary: Create a new outcome
      description: Open to all users.
      operationId: createOutcome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  outcomeStatus:
                    type: string
                    example: 'Open'
                  transactionType:
                    type: string
                    example: 'Call'
                  numberCalled:
                    type: string
                    example: '0123456789'
                  emailUsed:
                    type: string
                    example: 'pluto@example.com'
                  contactPerson:
                    type: string
                    example: 'Pluto'
                  outcomeResolution:
                    type: string
                    example: 'Customer requested a callback'
                  ptpDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  ptpAmount:
                    type: integer
                    example: 1500.00
                  debitResubmissionDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  debitResubmissionAmount:
                    type: integer
                    example: 2000.00
                  outcomeNotes:
                    type: string
                    example: 'Need to call back the customer'
                  nextSteps:
                    type: string
                    example: 'Need to keep pressing for payment'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  closedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  closedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_caseId:
                    type: integer
                    example: 123
              required:
                - createdBy
                - createdDate
                - f_caseId
      responses:
        '200':
          description: Outcome created successfully. The id of the new outcome is returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
        '400':
          description: Outcome id already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Outcome id '123' already exists"
        '401':
          $ref: '#/components/responses/UnauthorisedError'
  /outcomes/{id}:
    parameters:
      - in: path
        name: id
        description: Outcome id
        required: true
        example: 123
        schema:
          type: integer
    get:
      summary: Get a single outcome by outcome id
      description: Retrieve a single outcome by outcome id.
      operationId: getOutcomeById
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the specified customer
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  outcomeStatus:
                    type: string
                    example: 'Open'
                  transactionType:
                    type: string
                    example: 'Call'
                  numberCalled:
                    type: string
                    example: '0123456789'
                  emailUsed:
                    type: string
                    example: 'pluto@example.com'
                  contactPerson:
                    type: string
                    example: 'Pluto'
                  outcomeResolution:
                    type: string
                    example: 'Customer requested a callback'
                  ptpDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  ptpAmount:
                    type: integer
                    example: 1500.00
                  debitResubmissionDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  debitResubmissionAmount:
                    type: integer
                    example: 2000.00
                  outcomeNotes:
                    type: string
                    example: 'Need to call back the customer'
                  nextSteps:
                    type: string
                    example: 'Need to keep pressing for payment'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  closedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  closedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_caseId:
                    type: integer
                    example: 123
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    put:
      summary: Update an outcome
      description: Update details of an outcome.
      operationId: updateOutcome
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                outcomeStatus:
                  type: string
                  example: 'Open'
                transactionType:
                  type: string
                  example: 'Call'
                numberCalled:
                  type: string
                  example: '0123456789'
                emailUsed:
                  type: string
                  example: 'pluto@example.com'
                contactPerson:
                  type: string
                  example: 'Pluto'
                outcomeResolution:
                  type: string
                  example: 'Customer requested a callback'
                ptpDate:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                ptpAmount:
                  type: integer
                  example: 1500.00
                debitResubmissionDate:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                debitResubmissionAmount:
                  type: integer
                  example: 2000.00
                outcomeNotes:
                  type: string
                  example: 'Need to call back the customer'
                nextSteps:
                  type: string
                  example: 'Need to keep pressing for payment'
                createdDate:
                  type: string
                  example: '2020-05-05T09:12:57.848Z'
                createdBy:
                  type: string
                  example: 'mickey@example.com'
                closedDate:
                  type: string
                  example: '2020-05-08T03:11:21.553Z'
                closedBy:
                  type: string
                  example: 'mickey@example.com'
                f_caseId:
                  type: integer
                  example: 123
      responses:
        '200':
          description: Outcome updated successfully. The details of the updated outcome are returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  outcomeStatus:
                    type: string
                    example: 'Open'
                  transactionType:
                    type: string
                    example: 'Call'
                  numberCalled:
                    type: string
                    example: '0123456789'
                  emailUsed:
                    type: string
                    example: 'pluto@example.com'
                  contactPerson:
                    type: string
                    example: 'Pluto'
                  outcomeResolution:
                    type: string
                    example: 'Customer requested a callback'
                  ptpDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  ptpAmount:
                    type: integer
                    example: 1500.00
                  debitResubmissionDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  debitResubmissionAmount:
                    type: integer
                    example: 2000.00
                  outcomeNotes:
                    type: string
                    example: 'Need to call back the customer'
                  nextSteps:
                    type: string
                    example: 'Need to keep pressing for payment'
                  createdDate:
                    type: string
                    example: '2020-05-05T09:12:57.848Z'
                  createdBy:
                    type: string
                    example: 'mickey@example.com'
                  closedDate:
                    type: string
                    example: '2020-05-08T03:11:21.553Z'
                  closedBy:
                    type: string
                    example: 'mickey@example.com'
                  f_caseId:
                    type: integer
                    example: 123
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
    delete:
      summary: Delete an outcome
      description: Delete an outcome.
      operationId: deleteOutcome
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Outcome deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Outcome deleted successfully'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorisedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorisedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Unauthorised'
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Not Found'